# docs/swagger.yaml
definitions:
  models.Transaction:
    properties:
      id:
        type: string
      item_id:
        type: string
      payment_amount:
        type: number
      payment_time:
        type: string
      recipient_id:
        type: string
      sender_id:
        type: string
    type: object
host: localhost:9000
info:
  contact:
    email: dimash.amurzakov@gmail.com
    name: Dinmikhamed Amurzakov
  description: Microservice with transactions
  termsOfService: http://swagger.io/terms/
  title: Transactions microservice API
  version: "1.0"
paths:
  /trans/create:
    post:
      consumes:
      - application/json
      description: create transaction
      operationId: create-transaction
      parameters:
      - description: transaction info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            type: string
        "400":
          description: message
          schema:
            type: string
        "404":
          description: message
          schema:
            type: string
        "500":
          description: message
          schema:
            type: string
        default:
          description: message
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: CreateTrans
      tags:
      - transaction
  /trans/delete/{id}:
    get:
      consumes:
      - application/json
      description: delete transaction by id
      operationId: delete-transaction
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: message
          schema:
            type: string
        "404":
          description: message
          schema:
            type: string
        "500":
          description: message
          schema:
            type: string
        default:
          description: message
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: DeleteTrans
      tags:
      - transaction
  /trans/get_by_id/{id}:
    get:
      consumes:
      - application/json
      description: get transaction by ID
      operationId: get-transaction-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: message
          schema:
            type: string
        "404":
          description: message
          schema:
            type: string
        "500":
          description: message
          schema:
            type: string
        default:
          description: message
          schema:
            type: string
      summary: GetTransByID
      tags:
      - transaction
  /trans/list:
    get:
      consumes:
      - application/json
      description: list transactions
      operationId: list-transactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "400":
          description: message
          schema:
            type: string
        "404":
          description: message
          schema:
            type: string
        "500":
          description: message
          schema:
            type: string
        default:
          description: message
          schema:
            type: string
      summary: ListTrans
      tags:
      - transaction
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"