# internal/models/transaction.go
package models

import "time"

type Transaction struct {
	ID            string    `json:"id"`
	SenderID      string    `json:"sender_id"`
	RecipientID   string    `json:"recipient_id"`
	PaymentTime   time.Time `json:"payment_time"`
	ItemID        string    `json:"item_id"`
	PaymentAmount float64   `json:"payment_amount"`
}